[FIX] Move function to Utils
